feat: add manage vault in table
