+++ src/components/managedVaults/community/table/useCommunityVaultsColumns.tsx
@@ -2,16 +2,12 @@ import { ColumnDef, Row } from '@tanstack/react-table'
 import Apy, { APY_META } from 'components/earn/lend/Table/Columns/Apy'
 import Fee, { FEE_META } from 'components/managedVaults/common/table/columns/Fee'
 import Title, { TITLE_META } from 'components/managedVaults/common/table/columns/Title'
-import Details, { DETAILS_META } from 'components/managedVaults/community/table/column/Details'
 import MyPosition, {
   MY_POSITION_META,
 } from 'components/managedVaults/community/table/column/MyPosition'
 import TVL, { TVL_META } from 'components/managedVaults/community/table/column/TVL'
-import { useCallback, useMemo } from 'react'
-import { useNavigate, useSearchParams } from 'react-router-dom'
-import useStore from 'store'
-import { getRoute } from 'utils/route'
-import useVaultAssets from 'hooks/assets/useVaultAssets'
+import Manage, { MANAGE_META } from 'components/managedVaults/community/table/column/Manage'
+import { useMemo } from 'react'
 import { convertAprToApy } from 'utils/parsers'
 
 interface Props {
@@ -21,17 +17,6 @@ interface Props {
 
 export default function useCommunityVaultsColumns(props: Props) {
   const { isLoading, showPosition } = props
-  const address = useStore((s) => s.address)
-  const [searchParams] = useSearchParams()
-  const navigate = useNavigate()
-  const vaultAssets = useVaultAssets()
-
-  const handleVaultDetails = useCallback(
-    (vaultAddress: string) => {
-      navigate(getRoute(`vaults/${vaultAddress}/details` as Page, searchParams, address))
-    },
-    [address, navigate, searchParams],
-  )
 
   return useMemo<ColumnDef<ManagedVaultsData>[]>(() => {
     return [
@@ -75,14 +60,11 @@ export default function useCommunityVaultsColumns(props: Props) {
         ),
       },
       {
-        ...DETAILS_META,
+        ...MANAGE_META,
         cell: ({ row }: { row: Row<ManagedVaultsData> }) => (
-          <Details
-            isLoading={isLoading}
-            handleVaultDetails={() => handleVaultDetails(row.original.vault_address)}
-          />
+          <Manage isLoading={isLoading} vault={row.original} />
         ),
       },
     ]
-  }, [isLoading, handleVaultDetails, showPosition])
+  }, [isLoading, showPosition])
 }
